nlp/FinalReport/requirement.txt
matplotlib
future
gensim
librosa
nltk
numpy
pandas
requests
regex
sklearn
seaborn
textdistance
pytorch
tqdm
wheel
